// src/loan-accounts/types/account-state-actions.ts
export enum AccountStateActions {
  REQUEST_APPROVAL = 'REQUEST_APPROVAL',
  SET_INCOMPLETE = 'SET_INCOMPLETE',
  REJECT = 'REJECT',
  WITHDRAW = 'WITHDRAW',
  APPROVE = 'APPROVE',
  UNDO = 'UNDO',
  DISBURSE = 'DISBURSE',
  WRITE_OFF = 'WRITE_OFF',
  TERMINATE = 'TERMINATE',
  MAKE_REPAYMENT = 'MAKE_REPAYMENT',
  MAKE_INSTALLMENT_REPAYMENT = 'MAKE_INSTALLMENT_REPAYMENT',
  PAY_OFF = 'PAY_OFF',
  EARLY_PAYMENT = 'EARLY_PAYMENT',
}